+++ src/browser/index.jsx
@@ -417,6 +417,10 @@ var MattermostView = React.createClass({
           }
         });
       }
+
+      require('electron-context-menu')({
+        window: webview
+      });
     });
 
     webview.addEventListener('ipc-message', function(event) {
@@ -597,12 +601,6 @@ try {
 if (config.teams.length === 0) {
   window.location = 'settings.html';
 }
-
-var contextMenu = require('./menus/context');
-var menu = contextMenu.createDefault();
-window.addEventListener('contextmenu', function(e) {
-  menu.popup(remote.getCurrentWindow());
-}, false);
 
 var showUnreadBadgeWindows = function(unreadCount, mentionCount) {
   const badge = require('./js/badge');